+++ src/features/encoders/browserGIF/client/index.ts
@@ -1,4 +1,4 @@
-import { canvasEncodeTest, canvasEncode } from 'client/lazy-app/util';
+import { canvasEncodeTest, canvasEncode } from 'client/lazy-app/util/canvas';
 import WorkerBridge from 'client/lazy-app/worker-bridge';
 import { EncodeOptions, mimeType } from '../shared/meta';
